+++ docs/.vuepress/config.js
@@ -202,7 +202,9 @@ module.exports = {
             ["composable/web/fetch", "fetch"],
             ["composable/web/webSocket", "webSocket"],
             ["composable/web/intersectionObserver", "IntersectionObserver"],
-            ["composable/web/networkInformation", "NetworkInformation"]
+            ["composable/web/networkInformation", "NetworkInformation"],
+            ["composable/web/online", "Navigator.onLine"],
+            ["composable/web/pageVisibility", "PageVisibilityAPI"]
           ]
         },
         {