+++ packages/web/src/web/networkInformation.ts
@@ -1,5 +1,5 @@
-import { computed, Ref, ref, onUnmounted } from "@vue/composition-api";
-import { RemoveEventFunction } from "../event";
+import { computed, Ref, ref } from "@vue/composition-api";
+import { RemoveEventFunction, useEvent } from "../event";
 import { NO_OP } from "@vue-composable/core";
 
 interface NetworkInformationEventMap {
@@ -123,14 +123,14 @@ export function useNetworkInformation(): NetworkInformationReturn {
       type.value = connection.type;
     };
 
-    remove = () => {
-      connection.removeEventListener("change", handler);
-    };
+    remove = useEvent<NetworkInformation, NetworkInformationEventMap, "change">(
+      connection,
+      "change",
+      handler,
+      { passive: true }
+    );
 
-    connection.addEventListener("change", handler, { passive: true });
     handler();
-
-    onUnmounted(remove);
   } else if (__DEV__) {
     console.warn(
       "[navigator.connection] not found, networkInformation not available."